+++ src/firmwareanalysis/azext_firmwareanalysis/aaz/latest/firmwareanalysis/firmware/__init__.py
@@ -15,7 +15,6 @@ from ._crypto_certificate import *
 from ._crypto_key import *
 from ._cve import *
 from ._delete import *
-from ._generate_filesystem_download_url import *
 from ._list import *
 from ._password_hash import *
 from ._sbom_component import *