+++ src/firmwareanalysis/HISTORY.rst
@@ -6,6 +6,8 @@ Release History
 2.0.0
 +++++
 * Updating version for 2025-08-02 swagger release
+* Added workspace usage-metric command
+* Deprecated generate filesystem download url command
 
 1.0.0
 ++++++